requirements.txt
backports.functools-lru-cache==1.5
cycler==0.10.0
Django==2.2.6
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.15.4
Pillow==5.3.0
pycairo==1.18.0
PyGObject==3.30.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
subprocess32==3.5.3
